// src/cache/ecom-cache-key.enum.ts
export enum EcomCoreCacheKeyEnum {
  COUNTRIES = 'COUNTRIES',
  STATES = 'STATES',
  DISTRICTS = 'DISTRICTS',
  CURRENCIES = 'CURRENCIES',
  RESIDENTIAL_AREAS = 'RESIDENTIAL_AREAS',
  THANAS = 'THANAS',
}

export enum EcomCatelogCacheKeyEnum {
  TYPES_FIND_ALL = 'TYPES_FIND_ALL',
  TYPES_SHOPS_FOR_HOME = 'TYPES_SHOPS_FOR_HOME',
  CATEGORY_FIND_ALL = 'CATEGORY_FIND_ALL',
  CATEGORY_TREES = 'CATEGORY_TREES',
  PRODUCT_FIND_ALL = 'PRODUCT_FIND_ALL',
  CATEGORY_PRODUCTS_FOR_HOME = 'CATEGORY_PRODUCTS_FOR_HOME',
}